Moving PageLoadMetricsMemoryTrackerFactory to anonymous namespace
