+++ cobalt/browser/metrics/page_load_metrics_initialize.cc
@@ -14,8 +14,10 @@
 
 #include "cobalt/browser/metrics/page_load_metrics_initialize.h"
 
+#include "base/memory/singleton.h"
 #include "base/notreached.h"
-#include "cobalt/browser/metrics/page_load_metrics_memory_tracker_factory.h"
+#include "components/keyed_service/content/browser_context_dependency_manager.h"
+#include "components/keyed_service/content/browser_context_keyed_service_factory.h"
 #include "components/page_load_metrics/browser/metrics_web_contents_observer.h"
 #include "components/page_load_metrics/browser/page_load_metrics_embedder_base.h"
 #include "components/page_load_metrics/browser/page_load_metrics_memory_tracker.h"
@@ -24,13 +26,64 @@ namespace content {
 class BrowserContext;
 }  // namespace content
 
-namespace page_load_metrics {
-class PageLoadMetricsMemoryTracker;
-}  // namespace page_load_metrics
-
 namespace cobalt {
 
 namespace {
+
+class PageLoadMetricsMemoryTrackerFactory
+    : public BrowserContextKeyedServiceFactory {
+ public:
+  static page_load_metrics::PageLoadMetricsMemoryTracker* GetForBrowserContext(
+      content::BrowserContext* context);
+
+  static PageLoadMetricsMemoryTrackerFactory* GetInstance();
+
+  PageLoadMetricsMemoryTrackerFactory();
+
+ private:
+  // BrowserContextKeyedServiceFactory:
+  bool ServiceIsCreatedWithBrowserContext() const override;
+
+  std::unique_ptr<KeyedService> BuildServiceInstanceForBrowserContext(
+      content::BrowserContext* context) const override;
+
+  content::BrowserContext* GetBrowserContextToUse(
+      content::BrowserContext* context) const override;
+};
+
+page_load_metrics::PageLoadMetricsMemoryTracker*
+PageLoadMetricsMemoryTrackerFactory::GetForBrowserContext(
+    content::BrowserContext* context) {
+  return static_cast<page_load_metrics::PageLoadMetricsMemoryTracker*>(
+      GetInstance()->GetServiceForBrowserContext(context, /*create=*/true));
+}
+
+PageLoadMetricsMemoryTrackerFactory*
+PageLoadMetricsMemoryTrackerFactory::GetInstance() {
+  return base::Singleton<PageLoadMetricsMemoryTrackerFactory>::get();
+}
+
+PageLoadMetricsMemoryTrackerFactory::PageLoadMetricsMemoryTrackerFactory()
+    : BrowserContextKeyedServiceFactory(
+          "PageLoadMetricsMemoryTracker",
+          BrowserContextDependencyManager::GetInstance()) {}
+
+bool PageLoadMetricsMemoryTrackerFactory::ServiceIsCreatedWithBrowserContext()
+    const {
+  return base::FeatureList::IsEnabled(features::kV8PerFrameMemoryMonitoring);
+}
+
+std::unique_ptr<KeyedService>
+PageLoadMetricsMemoryTrackerFactory::BuildServiceInstanceForBrowserContext(
+    content::BrowserContext* context) const {
+  return std::make_unique<page_load_metrics::PageLoadMetricsMemoryTracker>();
+}
+
+content::BrowserContext*
+PageLoadMetricsMemoryTrackerFactory::GetBrowserContextToUse(
+    content::BrowserContext* context) const {
+  return context;
+}
 
 class PageLoadMetricsEmbedder
     : public page_load_metrics::PageLoadMetricsEmbedderBase {
@@ -75,8 +128,8 @@ void PageLoadMetricsEmbedder::RegisterEmbedderObservers(
 page_load_metrics::PageLoadMetricsMemoryTracker*
 PageLoadMetricsEmbedder::GetMemoryTrackerForBrowserContext(
     content::BrowserContext* browser_context) {
-  return page_load_metrics::PageLoadMetricsMemoryTrackerFactory::
-      GetForBrowserContext(browser_context);
+  return PageLoadMetricsMemoryTrackerFactory::GetForBrowserContext(
+      browser_context);
 }
 
 }  // namespace
